% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cookies.R
\name{cookies}
\alias{cookies}
\title{Derive radially constrained spatial subsamples from occurrence data}
\usage{
cookies(
  dat,
  xy,
  r,
  rarefy = TRUE,
  weight = FALSE,
  seeding = NULL,
  iter = 100,
  nSite = 3,
  crs = "EPSG:8857",
  output = "locs"
)
}
\arguments{
\item{dat}{A data.frame or matrix containing taxon names, Cartesian coordinates, and any associated variables.}

\item{xy}{A vector of two elements, specifying the name of columns in \code{dat} containing Cartesian coordinates, e.g. longitude and latitude. Coordinates for any shared sampling sites should be identical, and where sites are raster cells, coordinates are usually expected to be cell centroids.}

\item{r}{A numeric value specifying the radius (in metres) to use to define radially constrained regions.}

\item{rarefy}{Whether sites within radially constrained regions should be subsampled (\code{rarefy = TRUE}, default) or not (\code{rarefy = FALSE}). If \code{TRUE}, \code{nSite} sites will be drawn from \code{iter} randomly selected radially constrained regions.}

\item{weight}{Whether sites within the radially constrained region should be drawn at random (\code{weight = FALSE}, default) or with probability inversely proportional to the square of their distance from the centre of the region (\code{weight = TRUE}) during subsampling. Only applicable when \code{rarefy = TRUE}.}

\item{seeding}{Either \code{NULL} (the default) or a \code{data.frame} with additional class \code{seedMatrix} generated using the \code{\link[=defineSeeds]{defineSeeds()}} or \code{cookies} (with \code{output = "seeds"}) functions, specifying the location of the centroids to be used to define radially constrained regions. If \code{NULL}, each unique populated site within \code{dat} will be assessed for viability as the center of a radially constrained region.}

\item{iter}{A numeric value specifying the number of subsamples to be drawn from each radially constrained region if \code{rarefy = "divvySites"}, \code{"weightedDivvySites"}, \code{"sites"}, \code{"occs"}, or \code{"sitesThenOccs"}. Default is \code{100}.}

\item{nSite}{A numeric value specifying the minimum number of unique populated sites that must fall within a radially constrained region for it to be considered viable. Also, specifies the number of sites that are randomly drawn \emph{without} replacement from each radially constrained region if \code{rarefy = "sites"} or \code{"sitesThenOccs"}. Default is \code{3}.}

\item{crs}{A character string specifying a coordinate reference system (CRS). You can use the following formats to define coordinate reference systems: WKT, PROJ.4 (e.g., \code{crs = +proj=longlat +datum=WGS84}), or an EPSG code (e.g., \code{crs = "EPSG:4326"}). But note that the PROJ.4 notation has been deprecated, and you can only use it with the WGS84/NAD83 and NAD27 datums. Other datums are silently ignored. Default is \code{"EPSG:8857"}.}

\item{output}{A character string, either \code{"full"},\code{"locs"}, or \code{"seeds"}. Specifies whether the returned data should be two columns of subsample site coordinates (\code{output = "locs"}), the subset of rows from dat associated with those coordinates (\code{output = "full"}), or a data.frame of the seeds used (\code{output = "seeds"}).}
}
\value{
A list of length iter.
\itemize{
\item If \code{rarefy = TRUE}, each element is a data.frame or matrix (matching the class of \code{dat}) with \emph{l} rows, where \emph{l} is the number of occurrences associated with \code{nSite} sites within a viable radially constrained region. The radially constrained region of origin for each subsample is chosen at random.
\item If \code{rarefy = FALSE}, each element is a data.frame or matrix (matching the class of \code{dat}) containing \emph{all} occurrences associated with \emph{all} sites within \emph{each} viable radially constrained region.
If \code{output = 'locs'} (default), only the coordinates of the \code{dat} rows associated with radially constrained regions/subsamples are returned. If \code{output = 'full'}, all \code{dat} columns are returned for the rows associated with the radially constrained region/subsample. If \code{output = 'seeds'}, only the coordinates of the viable seeds are returned.
}
}
\description{
This function is an expanded version of \code{\link[divvy:cookies]{divvy::cookies()}}.

In brief, the function takes a single location as a starting (seed) point, circumscribes a circular buffer of \code{r} metres around it, and then identifies which observations fall within this circle. This process is repeated for each unique site in the dataset (unless these seed points are manually specified). The radially constrained regions (RCRs) that result are then assessed against user-specified criteria; if they contain insufficient data (measured by the number of unique populated sites they contain) and/or overlap with other RCRs by more than is deemed acceptable, they are dropped from the analysis. Those that are retained are then either returned in full (i.e., all occurrences that fall within them are returned) or subsampled.

This version of the function differs from the original version in two key ways:
\enumerate{
\item Allows users to explore viability of specific regions for spatial subsampling through manual seeding.
\item Allows users to assess how changing values of \code{r}, \code{nSite}, and \code{oThreshold} affects the number of viable RCRs through the new option to return them without subsampling.
\item It allows users to return the viable seed points as a data.frame with additional class \code{seedMatrix}, which can be used to manually seed (via argument \code{seeding}) future runs.
}

The function also now returns the coordinates of the seed point in each subsample when \code{output = "full"}, which can be used to specify the location of radially constrained region from which the subsample is drawn as a random effect in mixed models.
}
\section{Manual seeding and raw radially constrained region outputs}{
These functionalities were added to give the function greater exploratory utility. Now users can repeatedly sample the same regions using different radii (\code{r}) and site minima (\code{nSite}) to determine optimum values for data sampling.
}

\section{Subsampling procedures}{
For explanations of how the subsampling procedures used when \code{rarefy = TRUE} and \code{weight = TRUE/FALSE}, see \code{\link[divvy:cookies]{divvy::cookies()}} documentation.
}

\section{Returning seeds}{
This functionality was added to further empower users to explore their data.
}

\examples{
# First, load terra
library(terra)
# Generate occurrence data
n <- 100
set.seed(5)
# 100 sets of x and y coordinates
x <- runif(n, 0, 50)
y <- runif(n, 0, 50)
# Units in equal earth are in meters
# So if we consider x and y as given in km
x <- x * 1000
y <- y * 1000
# Set name for coords
xy.name <- c("x", "y")
# Combine into data frame and label columns
pts <- data.frame(x, y)
colnames(pts) <- xy.name
# Try spatial subsampling
# 10km radius, minimum 3 sites, no overlap
subsamples <- cookies(pts, xy = xy.name,
r = 10000, rarefy = TRUE, output = "full")
}
