% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defineSeeds.R
\name{defineSeeds}
\alias{defineSeeds}
\title{Define centroids of radially constrained regions to be subsampled by \code{\link[=cookies]{cookies()}}}
\usage{
defineSeeds(dat, xy = c("x", "y"))
}
\arguments{
\item{dat}{A data.frame with at least 2 columns. These 2 columns must contain the Cartesian coordinates of the seeds to be used by \code{\link[=cookies]{cookies()}}. Their names must be submitted via argument 'xy' as a character vector.}

\item{xy}{A character vector with two elements, specifying the names of the columns in 'dat' containing the Cartesian coordinates of the seeds.}
}
\value{
A data frame.
}
\description{
This function defines seed points for use with \code{\link[=cookies2]{cookies2()}} (via argument \code{seeding}).
}
\examples{
# First, load terra
library(terra)
# Generate occurrence data
n <- 100
set.seed(5)
# 100 sets of x and y coordinates in meters
x <- runif(n, 0, 50000)
y <- runif(n, 0, 50000)
# Set name for coords
xy.name <- c("x", "y")
# Combine into data frame and label columns
pts <- data.frame(x, y)
colnames(pts) <- xy.name
# Toy dataset ready
# Set coordinates for our seeds
seedX <- c(10000, 15000, 20000, 25000, 30000, 35000, 40000)
seedY <- c(10000, 15000, 20000, 25000, 30000, 35000, 40000)
# Concatenate into data.frame with
# same coordinate names
dat <- data.frame(seedX, seedY)
colnames(dat) <- xy.name
# Define seedMatrix
seedMatrix <- defineSeeds(dat = dat, xy = xy.name)
# Get subsamples
subsamples <- cookies(dat = pts, xy = xy.name, r = 10000,
seeding = seedMatrix, output = "full")
}
