% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rasterOccData.R
\name{rasterOccData}
\alias{rasterOccData}
\title{Rasterise occurrence data and create seed matrix for cookies2}
\usage{
rasterOccData(
  occData,
  res,
  xyCoords1 = NULL,
  xyCoords2 = NULL,
  occData.crs = "EPSG:4326",
  raster.crs = "EPSG:8857",
  xyCoords = c("paleolng", "paleolat"),
  xyCell = c("cellX", "cellY"),
  uniqID = "cell"
)
}
\arguments{
\item{occData}{D}

\item{res}{D}

\item{xyCoords1}{D}

\item{xyCoords2}{D}

\item{occData.crs}{D}

\item{raster.crs}{D}

\item{xyCoords}{D}

\item{xyCell}{D}

\item{uniqID}{D}
}
\value{
D
}
\description{
Rasterise occurrence data and create seed matrix for cookies2
}
\examples{
print("example")
}
