% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cookies2.R
\name{cookies2}
\alias{cookies2}
\title{Derive radially constrained spatial subsamples from occurrence data}
\usage{
cookies2(
  dataMat,
  xy,
  r,
  seeding = NULL,
  rarefaction = "divvySites",
  reps = 100,
  nSites = 3,
  nOccs = 100,
  oThreshold = 0,
  oType = "sites",
  oPruningMode = "maxOccs",
  returnSeeds = F,
  crs = "EPSG:4326",
  output = "locs"
)
}
\arguments{
\item{dataMat}{D}

\item{xy}{D}

\item{r}{D}

\item{seeding}{D}

\item{rarefaction}{D}

\item{reps}{D}

\item{nSites}{D}

\item{nOccs}{D}

\item{oThreshold}{D}

\item{oType}{D}

\item{oPruningMode}{D}

\item{returnSeeds}{D}

\item{crs}{D}

\item{output}{D}
}
\value{
D
}
\description{
Derive radially constrained spatial subsamples from occurrence data
}
\examples{
print("example")
}
