% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defineSeeds.R
\name{defineSeeds}
\alias{defineSeeds}
\title{Define centroids of radially constrained regions to be subsampled by \code{\link[=cookies]{cookies()}}}
\usage{
defineSeeds(dat, xy = c("x", "y"))
}
\arguments{
\item{dat}{A data.frame with at least 2 columns. These 2 columns must contain the Cartesian coordinates of the seeds to be used by \code{\link[=cookies]{cookies()}}. Their names must be submitted via argument 'xy' as a character vector.}

\item{xy}{A character vector with two elements, specifying the names of the columns in 'dat' containing the Cartesian coordinates of the seeds.}
}
\value{
A data frame.
}
\description{
This function defines seed points for use with \code{\link[=cookies2]{cookies2()}} (via argument \code{seeding}).
}
