% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defineSeeds.R
\name{defineSeeds}
\alias{defineSeeds}
\title{Define centroids of radially constrained regions to be subsampled by \code{\link[=cookies]{cookies()}}}
\usage{
defineSeeds(grid, dat, xy = c("x", "y"))
}
\arguments{
\item{grid}{A SpatVector defining the gridding template to be used. Note, if terra::values(grid) = NA for all cells, the grid cells defined by the template will be numbered, starting from 1, and the resulting data frame of seeds will use this numbering scheme. It is \strong{highly recommended} that users label their gridding template SpatVectors before generating their seed matrices.}

\item{dat}{A data.frame with at least 2 columns. These 2 columns must contain the Cartesian coordinates of the seeds to be used by \code{\link[=cookies]{cookies()}}. Their names must be submitted via argument 'xy' as a character vector.}

\item{xy}{A character vector with two elements, specifying the names of the columns in 'dat' containing the Cartesian coordinates of the seeds.}
}
\value{
A data frame.
}
\description{
This function defines rasterised seed points for use with \code{\link[=cookies2]{cookies2()}} (via argument \code{seeding}).

It is \strong{highly recommended} that you first rasterise your occurrence data, and then use the same grid cell template you used to do so to define your seed matrix \strong{as this will ensure your seeds and point occurrences are rasterised using the same grid of cells}.
}
\examples{
# First, load terra
library(terra)
# Work in Equal Earth project coordinates
prj <- 'EPSG:8857'
# Now we generate occurrence data
n <- 100
set.seed(5)
# 100 sets of x and y coordinates in North Africa
x <- runif(n,  0, 30)
y <- runif(n, 10, 30)
# Units in equal earth are in meters
# So if we consider x and y as given in km
x <- x * 1000
y <- y * 1000
# Set name for coords
xy.name <- c("x", "y")
# Combine into data frame and label columns
pts <- data.frame(x, y)
colnames(pts) <- xy.name
# Finally, convert to SpatVector
ptsPrj <- terra::vect(pts, geom = xy.name,
                      crs = prj)
# Now let's define grid template with 5km grid cells
grid <- terra::project(x = terra::rast(),
                       y = prj, res = 5*1000)
# Number grid cells
terra::values(grid) <- 1:terra::ncell(grid)
# Now assign occurrences in toy dataset to grid cells
# First the cell number
pts[,"cell"] <- terra::cells(grid, ptsPrj)[,"cell"]
# Then the cell coordinates
pts[,c("cellX", "cellY")] <- terra::xyFromCell(grid,
pts[,"cell"])
# Toy dataset ready
# Now let's define out seeds in km
seedX <- c(155000, 160000, 165000)
seedY <- c(-5000, -10000, -15000)
# Concatenate into data.frame
dat <- data.frame(seedX, seedY)
colnames(dat) <- xy.name
# Use defineSeeds to create seed matrix
seedMatrix <- defineSeeds(grid = grid,
dat = dat, xy = xy.name)
# Use cookies to assess viability of 20km radius regions
}
